+++ src/commands.ts
@@ -48,6 +48,10 @@ import { SearchLogByText } from "./commands/search_log_by_text";
 import { Merge } from "./commands/merge";
 import { GenerateCommitMessage } from "./commands/generateCommitMessage";
 import { SelectAIModel } from "./commands/selectAIModel";
+import { Stage } from "./commands/stage";
+import { Unstage } from "./commands/unstage";
+import { StageAll } from "./commands/stageAll";
+import { UnstageAll } from "./commands/unstageAll";
 
 export function registerCommands(
   sourceControlManager: SourceControlManager,
@@ -101,4 +105,8 @@ export function registerCommands(
   disposables.push(new SearchLogByText());
   disposables.push(new GenerateCommitMessage());
   disposables.push(new SelectAIModel());
+  disposables.push(new Stage());
+  disposables.push(new Unstage());
+  disposables.push(new StageAll());
+  disposables.push(new UnstageAll());
 }